requirements.txt
astroid==2.4.0
attrs==16.3.0
backoff==1.8.0
certifi==2020.4.5.1
chardet==3.0.4
ciso8601==2.1.3
entrypoints==0.3
idna==2.8
isort==4.3.21
jsonschema==2.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pendulum==1.2.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.5.0
python-dateutil==2.8.1
pytz==2018.4
pytzdata==2019.3
requests==2.22.0
simplejson==3.11.1
singer-python==2.1.4
singer-tools==0.4.1
six==1.14.0
strict-rfc3339==0.7
tap-persistiq==0.0.2
terminaltables==3.1.0
toml==0.10.0
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.9
wrapt==1.12.1
